LPS-97280 Extract logic from model impl to AssetListHelper
